Tidy all reusable, non-trivial GUI related into common module
Re #170
