+++ mantidimaging/core/filters/circular_mask/circular_mask_gui.py
@@ -6,7 +6,7 @@ from . import execute
 
 
 def _gui_register(form):
-    from mantidimaging.gui.filters_window import add_property_to_form
+    from mantidimaging.gui.utility import add_property_to_form
 
     _, radius_field = add_property_to_form(
             'Radius', 'float', 0.95, (0.0, 1.0), form=form)
