+++ mantidimaging/core/filters/clip_values/clip_values_gui.py
@@ -6,7 +6,7 @@ from . import execute
 
 
 def _gui_register(form):
-    from mantidimaging.gui.filters_window import add_property_to_form
+    from mantidimaging.gui.utility import add_property_to_form
 
     value_range = (-10000000, 10000000)
