+++ mantidimaging/core/filters/crop_coords/crop_coords_gui.py
@@ -8,7 +8,7 @@ from . import crop_coords
 
 
 def _gui_register(form):
-    from mantidimaging.gui.filters_window import add_property_to_form
+    from mantidimaging.gui.utility import add_property_to_form
 
     add_property_to_form(
             'Select ROI on stack visualiser.', 'label', form=form)
