+++ mantidimaging/core/filters/stripe_removal/stripe_removal_gui.py
@@ -4,7 +4,7 @@ from . import execute, wavelet_names
 
 
 def _gui_register(form):
-    from mantidimaging.gui.filters_window import add_property_to_form
+    from mantidimaging.gui.utility import add_property_to_form
 
     # Filter type option
     _, value_filter_type = add_property_to_form(
