+++ mantidimaging/gui/filters_window/__init__.py
@@ -1,6 +1,5 @@
 from __future__ import absolute_import, division, print_function
 
-from .misc import add_property_to_form  # noqa: F401
 from .model import FiltersWindowModel  # noqa: F401
 from .presenter import FiltersWindowPresenter  # noqa: F401
 from .view import FiltersWindowView  # noqa: F401
