+++ mantidimaging/gui/filters_window/model.py
@@ -6,14 +6,12 @@ from logging import getLogger
 import numpy as np
 
 from mantidimaging.gui.stack_visualiser import Notification as SVNotification
-
+from mantidimaging.gui.utility import get_auto_params_from_stack
 from mantidimaging.core.utility.registrator import (
         get_package_children,
         import_items,
         register_into
     )
-
-from .misc import get_auto_params_from_stack
 
 
 def ensure_tuple(val):
