+++ mantidimaging/gui/utility/__init__.py
@@ -1,6 +1,12 @@
 from __future__ import absolute_import, division, print_function
 
-from .block_qt_signals import BlockQtSignals  # noqa: F401
-from .compile_ui import compile_ui  # noqa: F401
+from .get_auto_params_from_stack import get_auto_params_from_stack  # noqa: F401
+
+from .qt_helpers import (  # noqa: F401
+        compile_ui,
+        add_property_to_form,
+        delete_all_widgets_from_layout,
+        BlockQtSignals
+    )
 
 del absolute_import, division, print_function
